requirements.txt
flask
pyspellchecker
PyPDF2
chardet
gingerit
